Finalize chatbot name
