+++ config/checkstyle/suppressions.xml
@@ -7,5 +7,5 @@
 <suppressions>
     <suppress checks="JavadocType" files=".*Test\.java"/>
     <suppress checks="MissingJavadocMethodCheck" files=".*Test\.java"/>
-    <suppress checks="MissingJavadocMethodCheck" files="../src/main/lolok/exception/*.java"/>
+    <suppress checks="MissingJavadocMethodCheck" files="../src/main/loklok/exception/*.java"/>
 </suppressions>
