+++ src/main/java/loklok/exception/LokLokException.java
@@ -1,0 +1,21 @@
+package loklok.exception;
+
+/**
+ * An exception related to errors caused by chatbot.
+ */
+public class LokLokException extends RuntimeException {
+
+    /**
+     * Creates an exception object related to the chatbot.
+     *
+     * @param message The error message of the exception.
+     */
+    public LokLokException(String message) {
+        super(message);
+    }
+
+    @Override
+    public String toString() {
+        return "LokLokException: " + getMessage();
+    }
+}
